FullMatrix has been swapped by SimpleMatrix over all local solutions matrix of u0 and v0, all local shape functions matrix \phi_u and \phi_v in femgl.
A type is discoveried in beta3, beta4 and beta5 cell contribution fucntions. A comma was written as dot.
The goos thing is FullMatirx could initial object with one double number. Though FullMatrix was initialted
incorrectly, An assignmnent of 0 was called later on FullMatrix., this call corrected the value of FullMatrix objs.

On SimpleMatrix, this will generate compiling time error. And it got fixed.

necessary modifications were done for sol/CMakeLists.txt femgl/CMakeLists.txt.